Add compiler and compiler version to test case for auto-detecting SDK version.
Now if a test case has a compiler and a compiler version stored/determined, it overrides the selection given by the submitter. In the CLI, if the submitter provides the compiler as SDK, it tells the mesa instance (and any resulting test cases) to determine their own version IGNORING the user-provided compiler version. This effectively means that the user-provided compiler is still absolutely essential when using the CLI.