Add ssh/https support and other niceties.
In lib (not yet in executable), can use https or ssh to access GitHub.
This is saved in the config, so it should be a "set it and forget it"
situation, but we might still want to provide options in the exectuable.

A silly check is now done to make sure that git-lfs is installed, and
checkouts will loudly fail if it is not detected.

In general, most cases where a bash execution is necessary now fail
loudly if they don't return a success code. Exceptions are compilations
since we want those errors to be reported to the testhub.