Fixes for MESA_DIR and updating mirror.
A bit of a kludge, but this now sets `MESA_DIR` before running tests and reverts it afterwards so that tests should still compile and run fine even if test cases do not explicitly set their own version of `MESA_DIR`.

Also we now remove the work directory before fetching to alleviate issues with the fetch process.